split popup.js
