+++ lib/html.js
@@ -1,0 +1,5 @@
+export function htmlToDoc(html) {
+  const doc = document.implementation.createHTMLDocument('resp');
+  doc.documentElement.innerHTML = html;
+  return doc;
+}
