+++ popup.html
@@ -41,6 +41,6 @@
     <div id="result" class="result"></div>
   </div>
 
-  <script src="popup.js"></script>
+  <script type="module" src="lib/init.js"></script>
 </body>
 </html>
